pyproject.toml
[tool.poetry]
name = "djangotemplate"
version = "0.1.0"
description = ""
authors = ["hirowat2 <hiro@estatcamp.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
click = "8.1.3"
python-decouple = "3.6"
requests = "2.28.1"
django = "5.1.3"
django-extensions = "3.2.3"
psycopg2-binary = "^2.9.10"
gunicorn = "20.1.0"
whitenoise = "6.2.0"
autopep8 = "2.0.0"
isort = "5.10.1"
djhtml = "1.5.2"
django-widget-tweaks = "^1.5.0"
six = "1.16.0"
openpyxl = "^3.1.5"
rich = "12.6.0"
django-seed = "0.3.1"
pillow = "9.3.0"
ipdb = "^0.13.13"
ruff = "0.0.254"
faker-commerce = "1.0.3"
django-import-export = "3.1.0"
dask = "2023.3.1"
pyexcelerate = "0.10.0"
graphviz = "^0.20.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-dynamic-models = "^0.0.5"
pydotplus = "^2.0.2"
cookiecutter = "^2.6.0"
django-ninja = "^1.3.0"
django-cruds-adminlte = "^0.0.16"
django-crispy-forms = "^2.3"
django-grappelli = "^4.0.1"
django-jazzmin = "^3.0.1"
django-macros = "^0.4.0"
django-filter = "^24.3"
asteval = "^1.0.5"
django-dashboards = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
